Public API

Public API

Econometrics.absorb — Function.

absorb

Function for constructing the FunctionTerm{typeof(absorb)} used in decompose.

source

Econometrics.EconometricModel — Type.

EconometricModel(estimator::Type{<:Union{EconometricModel,ModelEstimator}},
                 f::FormulaTerm,
                 data;
                 contrasts::Dict{Symbol} = Dict{Symbol,Union{<:AbstractContrasts,<:AbstractTerm}}(),
                 wts::Union{Nothing,Symbol} = nothing,
                 panel::Union{Nothing,Symbol} = nothing,
                 time::Union{Nothing,Symbol} = nothing,
                 vce::VCE = OIM)

Formula has syntax:

@formula(response ~ exogenous + (endogenous ~ instruments) + absorb(highdimscontrols))

Data must implement the Tables.jl API and use CategoricalArrays (CategoricalVector)

Weights are taken as StatsBase.FrequencyWeights

Panel and time indicators are used for longitudinal estimators

Examples

model = fit(EconometricModel, formula, data, kwargs...)
model = fit(BetweenEstimator, formula, data, panel = :panel, kwargs...)
model = fit(RandomEffectsEstimator, formula, data, panel = :panel, time = :time, kwargs...)

source

Econometrics.BetweenEstimator — Type.

BetweenEstimator(effect::Symbol,
                 groups::Vector{Vector{Int}}) <: LinearModelEstimators

Continuous response estimator collapsing a dimension in a longitudinal setting.

source

Econometrics.RandomEffectsEstimator — Type.

RandomEffectsEstimator(pid::Tuple{Symbol,Vector{Vector{Int}}},
                       tid::Tuple{Symbol,Vector{Vector{Int}}},
                       idiosyncratic::Float64,
                       individual::Float64,
                       θ::Vector{Float64}) <: LinearModelEstimators

Swamy-Arora estimator.

source

Econometrics.VCE — Type.

VCE

Variance-covariance estimators.

Observed Information Matrix (OIM)

Heteroscedasticity Consistent: HC0, HC1, HC2, HC3, HC4

source