Interacting with a model

The API allows for:

build/estimate models

query models for statistics (e.g., goodness-of-fit statistics), components (parameter estimates)

export model estimates

Core Functionality

Econometrics — Module

Econometrics

Econometrics in Julia.

Econometrics.EconometricModel — Type

EconometricModel(estimator::Type{<:Union{EconometricModel,ModelEstimator}},
                 f::FormulaTerm,
                 data;
                 contrasts::Dict{Symbol} = Dict{Symbol,Union{<:AbstractContrasts,<:AbstractTerm}}(),
                 wts::Union{Nothing,Symbol} = nothing,
                 panel::Union{Nothing,Symbol} = nothing,
                 time::Union{Nothing,Symbol} = nothing,
                 vce::VCE = OIM)

Formula has syntax:

@formula(response ~ exogenous + (endogenous ~ instruments) + absorb(highdimscontrols))

Data must implement the Tables.jl API and use CategoricalArrays (CategoricalVector)

Weights are taken as StatsBase.FrequencyWeights

Panel and time indicators are used for longitudinal estimators

Examples

model = fit(EconometricModel, formula, data, kwargs...)
model = fit(BetweenEstimator, formula, data, panel = :panel, kwargs...)
model = fit(RandomEffectsEstimator, formula, data, panel = :panel, time = :time, kwargs...)

Statistical/Regression Model Abstraction

Missing docstring.

Missing docstring for aic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for aicc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for r2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for adjr2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for mss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for rss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for deviance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nulldeviance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for loglikelihood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nullloglikelihood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for coef. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dof. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dof_residual. Check Documenter's build log for details.

Missing docstring.

Missing docstring for coefnames(::EconometricModel). Check Documenter's build log for details.

StatsAPI.coeftable — Method

coeftable(obj::EconometricModel;
	  level::Real = 0.95)
coeftable(obj::EconometricModel{<:LinearModelEstimators};
	  level::Real = 0.95,
	  vce::VCE = obj.vce)

Return a table of class CoefTable with coefficients and related statistics. level determines the level for confidence intervals (by default, 95%). vce determines the variance-covariance estimator (by default, OIM).

Missing docstring.

Missing docstring for islinear. Check Documenter's build log for details.

Missing docstring.

Missing docstring for informationmatrix. Check Documenter's build log for details.

StatsAPI.vcov — Method

vcov(obj::EconometricModel)
vcov(obj::EconometricModel{<:LinearModelEstimators}, vce::VCE = obj.vce)

Return the variance-covariance matrix for the coefficients of the model. The vce argument allows to request variance estimators.

StatsAPI.stderror — Method

stderror(obj::EconometricModel)
stderror(obj::EconometricModel{<:LinearModelEstimators}, vce::VCE = obj.vce)

Return the standard errors for the coefficients of the model. The vce argument allows to request variance estimators.

StatsAPI.confint — Method

confint(obj::EconometricModel; se::AbstractVector{<:Real} = stderror(obj), level::Real = 0.95)

Compute the confidence intervals for coefficients, with confidence level level (by default, 95%). se can be provided as a precomputed value.

StatsModels.hasintercept — Function

hasintercept(obj::EconometricModel)::Bool

Return whether the model has an intercept.

Missing docstring.

Missing docstring for isfitted. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fit(::EconometricModel). Check Documenter's build log for details.

Missing docstring.

Missing docstring for fit!(::EconometricModel). Check Documenter's build log for details.

Missing docstring.

Missing docstring for response. Check Documenter's build log for details.

Missing docstring.

Missing docstring for meanresponse. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fitted. Check Documenter's build log for details.

Missing docstring.

Missing docstring for predict. Check Documenter's build log for details.

Missing docstring.

Missing docstring for modelmatrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for residuals. Check Documenter's build log for details.

Missing docstring.

Missing docstring for leverage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nobs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for weights(::EconometricModel). Check Documenter's build log for details.

Estimators

Econometrics.BetweenEstimator — Type

BetweenEstimator(effect::Symbol,
                 groups::Vector{Vector{Int}}) <: LinearModelEstimators

Continuous response estimator collapsing a dimension in a longitudinal setting.

Econometrics.RandomEffectsEstimator — Type

RandomEffectsEstimator(pid::Tuple{Symbol,Vector{Vector{Int}}},
                       tid::Tuple{Symbol,Vector{Vector{Int}}},
                       idiosyncratic::Float64,
                       individual::Float64,
                       θ::Vector{Float64}) <: LinearModelEstimators

Swamy-Arora estimator.

Econometrics.VCE — Type

VCE

Variance-covariance estimators.

Observed Information Matrix (OIM)

Heteroscedasticity Consistent: HC0, HC1, HC2, HC3, HC4

Formula Components

Econometrics.absorb — Function

absorb

Function for constructing the FunctionTerm{typeof(absorb)} used in decompose.