API

Interacting with a model

The API allows for:

build/estimate models

query models for statistics (e.g., goodness-of-fit statistics), components (parameter estimates)

export model estimates

Core Functionality

Econometrics — Module.

Econometrics

Econometrics in Julia.

source

Econometrics.EconometricModel — Type.

EconometricModel(estimator::Type{<:Union{EconometricModel,ModelEstimator}},
                 f::FormulaTerm,
                 data;
                 contrasts::Dict{Symbol} = Dict{Symbol,Union{<:AbstractContrasts,<:AbstractTerm}}(),
                 wts::Union{Nothing,Symbol} = nothing,
                 panel::Union{Nothing,Symbol} = nothing,
                 time::Union{Nothing,Symbol} = nothing,
                 vce::VCE = OIM)

Formula has syntax:

@formula(response ~ exogenous + (endogenous ~ instruments) + absorb(highdimscontrols))

Data must implement the Tables.jl API and use CategoricalArrays (CategoricalVector)

Weights are taken as StatsBase.FrequencyWeights

Panel and time indicators are used for longitudinal estimators

Examples

model = fit(EconometricModel, formula, data, kwargs...)
model = fit(BetweenEstimator, formula, data, panel = :panel, kwargs...)
model = fit(RandomEffectsEstimator, formula, data, panel = :panel, time = :time, kwargs...)

source

Statistical/Regression Model Abstraction

StatsBase.aic — Function.

aic(obj::StatisticalModel)

Akaike's Information Criterion, defined as $-2 \log L + 2k$, with $L$ the likelihood of the model, and k its number of consumed degrees of freedom (as returned by dof).

StatsBase.aicc — Function.

aicc(obj::StatisticalModel)

Corrected Akaike's Information Criterion for small sample sizes (Hurvich and Tsai 1989), defined as $-2 \log L + 2k + 2k(k-1)/(n-k-1)$, with $L$ the likelihood of the model, $k$ its number of consumed degrees of freedom (as returned by dof), and $n$ the number of observations (as returned by nobs).

StatsBase.bic — Function.

bic(obj::StatisticalModel)

Bayesian Information Criterion, defined as $-2 \log L + k \log n$, with $L$ the likelihood of the model, $k$ its number of consumed degrees of freedom (as returned by dof), and $n$ the number of observations (as returned by nobs).

StatsBase.r2 — Function.

r2(obj::StatisticalModel)
r²(obj::StatisticalModel)

Coefficient of determination (R-squared).

For a linear model, the R² is defined as $ESS/TSS$, with $ESS$ the explained sum of squares and $TSS$ the total sum of squares.

r2(obj::StatisticalModel, variant::Symbol)
r²(obj::StatisticalModel, variant::Symbol)

Pseudo-coefficient of determination (pseudo R-squared).

For nonlinear models, one of several pseudo R² definitions must be chosen via variant. Supported variants are:

:MacFadden (a.k.a. likelihood ratio index), defined as $1 - \log (L)/\log (L_0)$;

:CoxSnell, defined as $1 - (L_0/L)^{2/n}$;

:Nagelkerke, defined as $(1 - (L_0/L)^{2/n})/(1 - L_0^{2/n})$.

In the above formulas, $L$ is the likelihood of the model, $L_0$ is the likelihood of the null model (the model with only an intercept), $n$ is the number of observations, $y_i$ are the responses, $\hat{y}_i$ are fitted values and $\bar{y}$ is the average response.

Cox and Snell's R² should match the classical R² for linear models.

StatsBase.adjr2 — Function.

adjr2(obj::StatisticalModel)
adjr²(obj::StatisticalModel)

Adjusted coefficient of determination (adjusted R-squared).

For linear models, the adjusted R² is defined as $1 - (1 - (1-R^2)(n-1)/(n-p))$, with $R^2$ the coefficient of determination, $n$ the number of observations, and $p$ the number of coefficients (including the intercept). This definition is generally known as the Wherry Formula I.

adjr2(obj::StatisticalModel, variant::Symbol)
adjr²(obj::StatisticalModel, variant::Symbol)

Adjusted pseudo-coefficient of determination (adjusted pseudo R-squared).

For nonlinear models, one of the several pseudo R² definitions must be chosen via variant. The only currently supported variant is :MacFadden, defined as $1 - (\log (L) - k)/\log (L0)$. In this formula, $L$ is the likelihood of the model, $L0$ that of the null model (the model including only the intercept). These two quantities are taken to be minus half deviance of the corresponding models. $k$ is the number of consumed degrees of freedom of the model (as returned by dof).

StatsBase.mss — Function.

mss(obj::StatisticalModel)

Return the model sum of squares.

StatsBase.rss — Function.

rss(obj::StatisticalModel)

Return the residual sum of squares.

StatsBase.deviance — Function.

deviance(obj::StatisticalModel)

Return the deviance of the model relative to a reference, which is usually when applicable the saturated model. It is equal, up to a constant, to $-2 \log L$, with $L$ the likelihood of the model.

StatsBase.nulldeviance — Function.

nulldeviance(obj::StatisticalModel)

Return the deviance of the null model, that is the one including only the intercept.

StatsBase.loglikelihood — Function.

loglikelihood(obj::StatisticalModel)

Return the log-likelihood of the model.

StatsBase.nullloglikelihood — Function.

loglikelihood(obj::StatisticalModel)

Return the log-likelihood of the null model corresponding to model obj. This is usually the model containing only the intercept.

StatsBase.coef — Function.

coef(obj::StatisticalModel)

Return the coefficients of the model.

StatsBase.dof — Function.

dof(obj::StatisticalModel)

Return the number of degrees of freedom consumed in the model, including when applicable the intercept and the distribution's dispersion parameter.

StatsBase.dof_residual — Function.

dof_residual(obj::RegressionModel)

Return the residual degrees of freedom of the model.

StatsBase.coefnames — Method.

coefnames(obj::StatisticalModel)

Return the names of the coefficients.

StatsBase.coeftable — Method.

coeftable(obj::EconometricModel;
	  level::Real = 0.95)
coeftable(obj::EconometricModel{<:LinearModelEstimators};
	  level::Real = 0.95,
	  vce::VCE = obj.vce)

Return a table of class CoefTable with coefficients and related statistics. level determines the level for confidence intervals (by default, 95%). vce determines the variance-covariance estimator (by default, OIM).

source

StatsBase.islinear — Function.

islinear(obj::StatisticalModel)

Indicate whether the model is linear.

StatsBase.informationmatrix — Function.

informationmatrix(model::StatisticalModel; expected::Bool = true)

Return the information matrix. By default the Fisher information matrix is returned, while the observed information matrix can be requested with expected = false.

StatsBase.vcov — Method.

vcov(obj::EconometricModel)
vcov(obj::EconometricModel{<:LinearModelEstimators}, vce::VCE = obj.vce)

Return the variance-covariance matrix for the coefficients of the model. The vce argument allows to request variance estimators.

source

StatsBase.stderror — Method.

stderror(obj::EconometricModel)
stderror(obj::EconometricModel{<:LinearModelEstimators}, vce::VCE = obj.vce)

Return the standard errors for the coefficients of the model. The vce argument allows to request variance estimators.

source

StatsBase.confint — Method.

confint(obj::EconometricModel; se::AbstractVector{<:Real} = stderror(obj), level::Real = 0.95)

Compute the confidence intervals for coefficients, with confidence level level (by default, 95%). se can be provided as a precomputed value.

source

StatsModels.hasintercept — Function.

hasintercept(obj::EconometricModel)::Bool

Return whether the model has an intercept.

source

StatsBase.isfitted — Function.

isfitted(obj::StatisticalModel)

Indicate whether the model has been fitted.

StatsBase.fit — Method.

Fit a statistical model.

StatsBase.fit! — Method.

Fit a statistical model in-place.

StatsBase.response — Function.

response(obj::RegressionModel)

Return the model response (a.k.a. the dependent variable).

StatsBase.meanresponse — Function.

meanresponse(obj::RegressionModel)

Return the mean of the response.

StatsBase.fitted — Function.

fitted(obj::RegressionModel)

Return the fitted values of the model.

StatsBase.predict — Function.

predict(obj::RegressionModel, [newX])

Form the predicted response of model obj. An object with new covariate values newX can be supplied, which should have the same type and structure as that used to fit obj; e.g. for a GLM it would generally be a DataFrame with the same variable names as the original predictors.

StatsBase.modelmatrix — Function.

modelmatrix(obj::RegressionModel)

Return the model matrix (a.k.a. the design matrix).

StatsBase.residuals — Function.

residuals(obj::RegressionModel)

Return the residuals of the model.

StatsBase.leverage — Function.

leverage(obj::RegressionModel)

Return the diagonal of the projection matrix.

StatsBase.nobs — Function.

nobs(obj::StatisticalModel)

Return the number of independent observations on which the model was fitted. Be careful when using this information, as the definition of an independent observation may vary depending on the model, on the format used to pass the data, on the sampling plan (if specified), etc.

StatsBase.weights — Method.

weights(obj::StatisticalModel)

Return the weights used in the model.

Estimators

Econometrics.BetweenEstimator — Type.

BetweenEstimator(effect::Symbol,
                 groups::Vector{Vector{Int}}) <: LinearModelEstimators

Continuous response estimator collapsing a dimension in a longitudinal setting.

source

Econometrics.RandomEffectsEstimator — Type.

RandomEffectsEstimator(pid::Tuple{Symbol,Vector{Vector{Int}}},
                       tid::Tuple{Symbol,Vector{Vector{Int}}},
                       idiosyncratic::Float64,
                       individual::Float64,
                       θ::Vector{Float64}) <: LinearModelEstimators

Swamy-Arora estimator.

source

Econometrics.VCE — Type.

VCE

Variance-covariance estimators.

Observed Information Matrix (OIM)

Heteroscedasticity Consistent: HC0, HC1, HC2, HC3, HC4

source

Formula Components

Econometrics.absorb — Function.

absorb

Function for constructing the FunctionTerm{typeof(absorb)} used in decompose.

source